평점과 좋/싫 분리
